UI and code adjustments, Isa attack sprites tmmrw
Trying to think of a staff basic attack minigame. When it is implemented tommorow it will use the sword minigame one because minigames take a whole day to make. Afterwards I'll make a focus blast skill for practice (and for fun) and then I'll steal the code from my other repos to make item selection